requirements.txt
pandas>=2.0.0
datasets>=2.14.0
huggingface_hub>=0.17.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

